@@ -1,4 +1,5 @@
-from typing import Any, Callable
+from collections.abc import Callable
+from typing import Any
 
 from irsim.lib.algorithm.generation import generate_polygon, random_generate_polygon
 from irsim.lib.algorithm.kinematics import (
